+++ res/output.txt
@@ -1,3 +1,3 @@
-Вчу чкцчузао щеор йрд фхузкхпн хежуча энщхузербюнпе.
-Ут жшйкч меэнщхузет, е фучус хецэнщхузет ц нцфурбмузетнкс ткцпурбпнъ цыктехнкз.
-Те вчус фулершо зцк, теьткс ц чепуиу ткжурбэуиу щеоре.
+Это тестовый файл для проверки работы шифровальщика.
+Он будет зашифрован, а потом расшифрован с использованием нескольких сценариев.
+На этом пожалуй все, начнем с такого небольшого файла.
